refactor(result-dao): extracted the result query into a separate DAO.
