+++ data/src/main/java/com/anwera97/data/AppDatabase.kt
@@ -2,10 +2,7 @@ package com.anwera97.data
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.anwera97.data.dao.CompanionDao
-import com.anwera97.data.dao.DebtorsDao
-import com.anwera97.data.dao.ExpenditureDao
-import com.anwera97.data.dao.TripDao
+import com.anwera97.data.dao.*
 import com.anwera97.data.entities.CompanionEntity
 import com.anwera97.data.entities.Debtors
 import com.anwera97.data.entities.Expenditure
@@ -22,4 +19,5 @@ abstract class AppDatabase : RoomDatabase() {
     abstract fun companionDao(): CompanionDao
     abstract fun expenditureDao(): ExpenditureDao
     abstract fun debtorsDao(): DebtorsDao
+    abstract fun resultDao(): ResultDao
 }
